moved tests to experimentalStyled()
