+++ packages/material-ui/src/styles/experimentalStyled.test.js
@@ -547,5 +547,47 @@ describe('experimentalStyled', () => {
       expect(getByTestId('with-classes').getAttribute('data-with-classes')).to.equal('true');
       expect(getByTestId('without-classes').getAttribute('data-with-classes')).to.equal('false');
     });
+
+    it('classes props should be correctly applied to root and slot elements', () => {
+      const Child = (props) => {
+        const { classes = {}, className, ...other } = props;
+      
+        return (
+          <div
+            data-testid="child"
+            className={`${classes.root} ${className} MuiChild-root`}
+            {...other}
+          />
+        );
+      };
+      
+      const ParentRoot = styled('div', {}, { name: 'MuiParent', slot: 'Root' })``;
+      const ParentSlot = styled(Child, {}, { name: 'MuiChild', slot: 'Slot' })``;
+      
+      const Parent = (props) => {
+        const { classes = {}, className, ...other } = props;
+      
+        return (
+          <ParentRoot
+            data-testid="parent"
+            className={`${classes.root} ${className} MuiParent-root`}
+            {...other}
+          >
+            {/* The classes prop here should not be blocked by the styled() utility */}
+            <ParentSlot classes={{ root: classes.slot }}></ParentSlot>
+          </ParentRoot>
+        );
+      };
+      
+      const { container } = render(<Parent classes={{ root: 'root', slot: 'slot' }} />);
+
+      expect(container.getElementsByClassName('MuiParent-root')[0]?.classList.contains('root')).to.equal(true)
+
+      // child has the correct class
+      expect(container.getElementsByClassName('MuiChild-root')[0]?.classList.contains('slot')).to.equal(true);
+
+      // none of the elements should have the classes attribute
+      expect(container.querySelectorAll('[classes]').length).to.equal(0);
+    })
   });
 });
