+++ test/e2e/index.test.ts
@@ -115,27 +115,4 @@ describe('e2e', () => {
       );
     });
   });
-
-  describe('styled-engine', () => {
-    it('should forward classes to slots', async () => {
-      await renderFixture('styled-engine/ClassesPropsTest');
-
-      // root has the correct class
-      expect(
-        await page.evaluate(() =>
-          document.getElementsByClassName('MuiParent-root')[0]?.classList.contains('root'),
-        ),
-      ).to.equal(true);
-
-      // child has the correct class
-      expect(
-        await page.evaluate(() =>
-          document.getElementsByClassName('MuiChild-root')[0]?.classList.contains('slot'),
-        ),
-      ).to.equal(true);
-
-      // none of the elements should have the classes attribute
-      expect(await page.evaluate(() => document.querySelectorAll('[classes]').length)).to.equal(0);
-    });
-  });
 });
